+++ templates/user/unsubscribe.html
@@ -10,9 +10,6 @@
         </p>
 
         <form method="POST" class="mt-4">
-          <input type="hidden" name="u" value="{{ user.id }}">
-          <input type="hidden" name="t" value="{{ verifier }}">
-
           <div class="field">
             <div class="control">
               <button class="button is-primary is-fullwidth">Unsubscribe</button>
